Fix Stack Exchange authentication handler
The payload from the SE API contains an `items` array which contains a
single user object that contains all the claims. As such, the
authentication handler should extract that user object already and run
the claims actions on that single user object.

Previously, the array was attempted to be parsed as an object (which
failed) and the built-in claims mapping did expect the result to contain
the raw payload (with the `items` array).

Also, although specifying a `key` is recommended for Stack Exchange to
receive higher request quotas, it is not a mandatory argument to pass in
the API.