Add Roslyn static analyzers
Add Roslyn static analyzers to find common coding mistakes so they don't have to be found manually by-eye in Pull Request reviews, and fix or suppress the violations the analyzers find.
Some rules have been disabled as they are either not a concern, or have already been extensively "broken" in the existing code and are non-trivial to remedy and are mostly only a question of style.
Move secrets for the sample app into application.json so they are not hard-coded, which also makes the analyzers happy.