Wait longer before expiry
The macOS machines seem to be a bit "slow" so it seems that the two initial generations where they should be the same end up different because it takes too long to complete. If this doesn't work, I'll just disable the test as it's redundant for the v6 release anyway.