Fix flaky test (#570)
* Fix flaky test

Try and fix flaky Apple provider test by not validating token lifetimes and increasing the delay slightly.

* Make the asserts easier to read

Swap the order so the asserts make more sense if the test fails.

* Wait longer before expiry

The macOS machines seem to be a bit "slow" so it seems that the two initial generations where they should be the same end up different because it takes too long to complete. If this doesn't work, I'll just disable the test as it's redundant for the v6 release anyway.