chore(EtsyTests): apply workaround into PostConfigure test
User side Post configure is later than provider side post configure. this causes the Scope not to be evaluated and the Mappings are not applyed automatically
HACK: add them all manually for the test postConfigure until decision is made this could be put in validate override or always consumer side?